this is a recommended system for digikala
program have two function   that have different model one is unsupervised for  predict similar product base on features, and other predict that for a user  this product  is recommended or not, notice in code we use two dataset. we tag mlp model in function 2 , if you want use this , please un tag, SMOTE part and scaler part in code.